extracted resource tree
